+++ R/ferx_predict.R
@@ -42,6 +42,8 @@ ferx_predict <- function(model, data = NULL, fit = NULL) {
     theta = fit_pieces$theta,
     omega_flat = fit_pieces$omega_flat,
     omega_dim = fit_pieces$omega_dim,
-    sigma = fit_pieces$sigma
+    sigma = fit_pieces$sigma,
+    omega_iov_flat = fit_pieces$omega_iov_flat,
+    omega_iov_dim = fit_pieces$omega_iov_dim
   )
 }