+++ R/ferx_predict_survival.R
@@ -72,6 +72,8 @@ ferx_predict_survival <- function(model, data = NULL, times, fit = NULL) {
     theta = fit_pieces$theta,
     omega_flat = fit_pieces$omega_flat,
     omega_dim = fit_pieces$omega_dim,
-    sigma = fit_pieces$sigma
+    sigma = fit_pieces$sigma,
+    omega_iov_flat = fit_pieces$omega_iov_flat,
+    omega_iov_dim = fit_pieces$omega_iov_dim
   )
 }